Update http-service-lambda to work with http-service version 0.4
Added a new `serve()` API that returns a Future so that users can choose
how to spawn it and don't need to update http-service-lambda in order
to benefit from newer async runtime.

On top of that, the original `run()` API now shares the underlying
Tokio runtime used by lambda_http instead of running two runtimes.